Migrate from perl to python

+++ MegaLTR.sh
@@ -11,7 +11,8 @@ LTR_HARVEST=$(pwd)/bin/LTR_HARVEST_parallel/LTR_HARVEST_parallel
 LTRretriever=$(pwd)/bin/LTR_retriever/LTR_retriever
 chmod 775 $LTRretriever #Give execulting permissions for LTRretriever
 RUN=$(pwd)/bin/RUN
-chmod 775 $RUN/usearch11.0.667_i86linux32 #Give execulting permissions
+# USEARCH removed (32-bit incompatible). Using VSEARCH instead.
+ #Give execulting permissions
 eval "$(conda shell.bash hook)"
 ############################################################
 # Set variables                                            #
@@ -154,15 +155,15 @@ conda activate MegaLTR
          echo
          printf "\tCheck the FASTA File format.\n"
          gunzip -c "$Fastafilepath" >$userpath/$process_id.fna
-         perl $RUN/checkfasta.pl $userpath/$process_id.fna ### Check the FASTA File format
+         python3 $RUN/checkfasta.py $userpath/$process_id.fna ### Check the FASTA File format
       elif [[ $Fastafilepath =~ \.zip$ ]]; then
          echo
          printf "\tCheck the FASTA File format.\n"
          gunzip -c "$Fastafilepath" >$userpath/$process_id.fna
-         perl $RUN/checkfasta.pl $userpath/$process_id.fna ### Check the FASTA File format
+         python3 $RUN/checkfasta.py $userpath/$process_id.fna ### Check the FASTA File format
       elif ([ $(stat -c%s "$Fastafilepath") -gt 500 ]); then
          printf "\tCheck the FASTA File format.\n"
-         perl $RUN/checkfasta.pl $Fastafilepath ### Check the FASTA File format
+         python3 $RUN/checkfasta.py $Fastafilepath ### Check the FASTA File format
          cp $Fastafilepath $userpath/$process_id.fna #### copy fasta file tRNA files
       else
          printf "\n\tCheck the FASTA File format.\n"
@@ -206,13 +207,22 @@ conda activate MegaLTR
          printf "\n\t$now51 \tLTR_HARVEST Started %s\n"
          mkdir -p $userpath/density
          densitypath=$userpath/density
-         conda config --show envs_dirs >$densitypath/condapath
-         sed -i '1d' $densitypath/condapath
-         sed -i 's/  - //g' $densitypath/condapath
-            for condapath in `less $densitypath/condapath`
-            do
-               perl $LTR_HARVEST -seq $userpath/$process_id.fna -threads $threads -size 1000000 -time 500 -gt $condapath/MegaLTR/bin/gt $minlenltr $maxlenltr $similar > /dev/null 2>/dev/null
-            done
+         # --- Find GenomeTools (gt) binary ---
+GT_BIN=$(command -v gt)
+if [ -z "$GT_BIN" ]; then
+  echo "ERROR: GenomeTools (gt) not found in PATH"
+  exit 1
+fi
+
+# --- Run LTR_HARVEST once ---
+perl $LTR_HARVEST \
+  -seq $userpath/$process_id.fna \
+  -threads $threads \
+  -size 1000000 \
+  -time 500 \
+  -gt $GT_BIN \
+  $minlenltr $maxlenltr $similar \
+  > /dev/null 2>/dev/null
          cat $FASTA/$process_id.fna.harvest.combine.scn  $FASTA/$process_id.fna.finder.combine.scn >$FASTA/$process_id.all.harvest.finder.combine
 
          now5="$(date)"
@@ -325,7 +335,8 @@ conda activate MegaLTR
          LTRFiles=$userpath/LTRFiles
          cd $LTRFiles
          split -n l/100 $Others/$process_id.ids.extract_seq
-         python3 $RUN/LTR_Seq_threads.py $userpath/$process_id.fna  $LTRFiles $Collected_Files $threads $RUN/extractseq-id-start-end.pl
+         python3 $RUN/LTR_Seq_threads.py $userpath/$process_id.fna  $LTRFiles $Collected_Files $threads $RUN/extractseq-id-start-end.py
+
          cp $ltrdigest/"$process_id"_pbs.fas $Collected_Files/$process_id.PBS.Sequence.fa
          cp $ltrdigest/"$process_id"_ppt.fas $Collected_Files/$process_id.PPT.Sequence.fa      
          sed  -i '1i LTR-RT id\tPseudomolecules/scaffolds\tLTR-RT start\tLTR-RT end\tLTR-RT length\tlLTR start\tlLTR end\tlLTR length\trLTR start\trLTR end\trLTR length\tlTSD start\tlTSD end\tlTSD sequence\trTSD start\trTSD end\trTSD sequence\tPPT start\tPPT end\tPPT motif\tStrand\tPPT offset\tPBS start\tPBS end\tStrand\ttRNA id\ttRNA motif\tPBS offset\ttRNA offset\tPBS/tRNA\t\tClass\tSuperfamily\tClade\tComplete\tStrand\tDomains' $Collected_Files/LTR_Table_TEsorter_Digest.tsv
@@ -335,9 +346,29 @@ conda activate MegaLTR
          USERCH= mkdir -p $userpath/USERCH
          USERCH=$userpath/USERCH
 		   cp $Collected_Files/LTR-RT_Sequence.fa $USERCH/LTR-RT_Sequence.fa
-		   $RUN/usearch11.0.667_i86linux32  -sortbylength $USERCH/LTR-RT_Sequence.fa --fastaout $USERCH/LTR-RT_Sequence_sorted.fa --log $USERCH/usearch.log
-		   $RUN/usearch11.0.667_i86linux32  -cluster_fast  $USERCH/LTR-RT_Sequence_sorted.fa --id 0.9 --centroids $USERCH/LTR-RTs_non-redundant.fa --uc $USERCH/result.uc -consout $USERCH/LTR-RTs_conses.fa -msaout $USERCH/aligned.fasta --log $USERCH/usearch2.log
-		   rm $USERCH/aligned.* 
+		  # --- Use VSEARCH instead of USEARCH (32-bit incompatible binary) ---
+VSEARCH_BIN="/home/asmaa/miniconda3/envs/vsearch_env/bin/vsearch"
+
+if [ ! -x "$VSEARCH_BIN" ]; then
+  echo "ERROR: vsearch binary not found or not executable: $VSEARCH_BIN"
+  exit 1
+fi
+
+# Sort sequences by length
+$VSEARCH_BIN --sortbylength $USERCH/LTR-RT_Sequence.fa \
+  --output $USERCH/LTR-RT_Sequence_sorted.fa \
+  --log $USERCH/vsearch_sort.log
+
+# Cluster sequences (90% identity) to build non-redundant library
+$VSEARCH_BIN --cluster_fast $USERCH/LTR-RT_Sequence_sorted.fa \
+  --id 0.90 \
+  --centroids $USERCH/LTR-RTs_non-redundant.fa \
+  --uc $USERCH/result.uc \
+  --consout $USERCH/LTR-RTs_conses.fa \
+  --msaout $USERCH/aligned.fasta \
+  --threads $threads \
+  --log $USERCH/vsearch_cluster.log
+        		   rm $USERCH/aligned.* 
 		   cp $USERCH/LTR-RTs_non-redundant.fa $Collected_Files/LTR-RTs_non-redundant_library.fasta
 		   now100="$(date)"
